Add audio fades to macro 1 and 10 for between services
Camera 17 is the announcement rotation, which is embedded with Stream BGM. We fade this up and the master down between services so we do not hear the audience having full conversations.

We fade up on macro 1 for the countdown, instead of at macro 2, because often the band will come in early before the countdown is over, and we need to be able to hear them.