Correct the master fader and remove crossfade in annc rot
The master record mix is the ExternalXLR input, not the master output.

AFV will turn on the audio for a video source when we switch to that camera. Tested with AnncRotation, we do not need to cross-fade the audio for that channel down; instead we are leaving it at the default audio level