2019 08 14 changes (#1)
* Add audio fades to macro 1 and 10 for between services

Camera 17 is the announcement rotation, which is embedded with Stream BGM. We fade this up and the master down between services so we do not hear the audience having full conversations.

We fade up on macro 1 for the countdown, instead of at macro 2, because often the band will come in early before the countdown is over, and we need to be able to hear them.

* Set macro 1 to preview camera 9 as the last step

Often, directors forget to preview something useful when they hit macro 2 to start the service. Preview with a decent and safe camera so it's not a disaster when it happens

* Correct the master fader and remove crossfade in annc rot

The master record mix is the ExternalXLR input, not the master output.

AFV will turn on the audio for a video source when we switch to that camera. Tested with AnncRotation, we do not need to cross-fade the audio for that channel down; instead we are leaving it at the default audio level

* Rename 0 to 'unity' and remove an annc rotation comment