Set macro 14 to play the set video
We were using macro 14 to play the 5th video on the VTR every time. We encountered a weekend that required 3 videos to be played back. This changes the macro to play the set video - the producer or director can change the video set on the VTR and make sure it is the right video to play before the director takes the macro and runs it.