Update lower title thirds macros to have longer sleep frames
We were seeing flashes of the previous videos when hitting lower title third macros which meant that the macro was keying on too quickly. Fix is to increase the number of frames after loading the video so we can guarantee that the lower-third is loaded in the VTR before keying it on.

Increases frames from 1 to 15