Disentangle dismiss/outro/reset macros
There was a lot of overlap in these; I don't think we ever audited them
in the context of resuming in-person gatherings where we'd have to use
both dismissal and outro video macros.

In this setup, the Dismiss macro does much of what Reset normally would
have, since we want most of those things at this point. Outro does
little more than roll that clip and transition to it. Reset does all
the normal functions, except in a way that won't be visible if Dismiss
has been run (i.e. mixing instead of dipping).

The only functionality loss is that the overflow aux will get a hard
cut to annc rotation in the middle of a dip transition away from the
outro video. We just don't have an available ME that is transitioning
into the in-building announcement rotation at that point. Not the worst
thing.