Add a macro to send Cam 4 to all room screens
This is to support private baptisms that should not be streamed but should still show on IMAG and room screens.

It's very similar in concept to the existing M19 for showing graphics on all room screens after a dismissal. But it's actually based on M11 (start baptism) except with all of the ME1-related commands removed.

This will allow the director to fire M9 (outro video) which cuts both video and audio to a VTR on ME1, followed by this new M18 to set the room screens to the desired sources. At the end M10 will clean up everything and reset to normal.