Change the normal level for the XLR input from unity to +6db
In a couple macros we step ("fade") the XLR input level from zero
(-inf) to unity, or the other direction.

For reasons, we want to start using +6dB instead of unity. After
recording and saving a macro to check the numerical value of +6dB, I
found it to be ~double of the unity value (this is probably backed by
an unsigned 16-bit integer).

So in order to get to +6dB I just doubled or halved the existing step
values. Except for the last one, which I capped at the max value of
65381.